# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "SkyLab test application API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "159.223.220.58:8080",
    "basePath": "/",
    "paths": {
        "/quiz/category": {
            "get": {
                "description": "Search all categories in database and return in JSON",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quiz"
                ],
                "summary": "Get all categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Category"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add category in database (JSON body) and return in JSON",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quiz"
                ],
                "summary": "Add category",
                "parameters": [
                    {
                        "description": "Request body example",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CategoryCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Category"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/quiz/question": {
            "post": {
                "description": "Add question and answers in database (JSON body) and return in JSON",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quiz"
                ],
                "summary": "Add question and answers",
                "parameters": [
                    {
                        "description": "Request body example",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.QuestionCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Question"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/quiz/question/{id}": {
            "get": {
                "description": "Search all questions by category ID in database and return in JSON",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quiz"
                ],
                "summary": "Get questions by category ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Question"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/quiz/questions/upload": {
            "post": {
                "description": "IUpload questions from JSON file to database",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Quiz"
                ],
                "summary": "Upload questions from JSON file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Choose JSON file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SucsessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Answer": {
            "type": "object",
            "properties": {
                "answer_text": {
                    "type": "string",
                    "example": "Text of answer item"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_true": {
                    "type": "boolean"
                },
                "question_id": {
                    "type": "integer"
                }
            }
        },
        "main.AnswerCreate": {
            "type": "object",
            "properties": {
                "answer_text": {
                    "type": "string",
                    "example": "Text of answer item"
                },
                "is_true": {
                    "type": "boolean"
                }
            }
        },
        "main.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Category name"
                }
            }
        },
        "main.CategoryCreate": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Category name"
                }
            }
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error message"
                }
            }
        },
        "main.Question": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Answer"
                    }
                },
                "category_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "question_text": {
                    "type": "string"
                }
            }
        },
        "main.QuestionCreate": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.AnswerCreate"
                    }
                },
                "category_id": {
                    "type": "integer"
                },
                "question_text": {
                    "type": "string",
                    "example": "Text of question"
                }
            }
        },
        "main.SucsessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Sucsess message"
                }
            }
        }
    }
}